# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Nika Web Services",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1"
        }
    ],
    "paths": {
        "/addresses": {
            "get": {
                "tags": [
                    "آدرس"
                ],
                "summary": "لیست آدرس ها بصورت صفحه بندی",
                "description": "",
                "operationId": "9f2baf910ef29838b3ca9503a3ca6232",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "شماره صفحه",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "example": 1
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "تعداد نمایش در هر صفحه",
                        "schema": {
                            "type": "number"
                        },
                        "example": 10
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "شناسه صاحب آدرس. فقط کاربری که مدیر و دسترسی مجاز را دارد میتواند بر اساس این فیلد جستجو کند",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "نام صاحب آدرس. فقط کاربری که مدیر و دسترسی مجاز را دارد میتواند بر اساس این فیلد جستجو کند",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone_number",
                        "in": "query",
                        "description": "شماره تلفن آدرس",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "نوع. Home:خانه Office:دفتر",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Home",
                                "Office"
                            ]
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "چیدمان. oldest:قدیمی ترین latest:جدیدترین",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "latest",
                            "enum": [
                                "oldest",
                                "latest"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "عملیات موفق",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "آدرس"
                ],
                "summary": "ثبت آدرس",
                "description": "کاربری که  مجوز مدیریت آدرس ها یا ایجاد آدرس را دارد میتواند با ارسال user_id برای دیگران هم آدرس ثبت کند.",
                "operationId": "8f39f6321e2b382380b373355045c670",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "address",
                                    "city_id",
                                    "postal_code",
                                    "phone_number",
                                    "type"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "شناسه کاربر",
                                        "type": "number"
                                    },
                                    "city_id": {
                                        "description": "شناسه شهر",
                                        "type": "number"
                                    },
                                    "address": {
                                        "description": "آدرس",
                                        "type": "string"
                                    },
                                    "postal_code": {
                                        "description": "کد پستی. بدون خط تیره مثال: 5717858731",
                                        "type": "number"
                                    },
                                    "phone_number": {
                                        "description": "شماره تلفن یا شماره موبایل. تفلن ثابت باید با کد شهر وارد شود بدون خط تیره مثال: 02162999977",
                                        "type": "number"
                                    },
                                    "type": {
                                        "description": "Home: خانه - Office: دفتر",
                                        "type": "string",
                                        "enum": [
                                            "Home",
                                            "Office"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "ثبت با موفقیت انجام شد",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/addresses/{id}": {
            "post": {
                "tags": [
                    "آدرس"
                ],
                "summary": "بروزرسانی آدرس",
                "description": "",
                "operationId": "ca1b5b8d45ba8f732d9592785944f460",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "شناسه آدرس",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "_method",
                                    "address",
                                    "city_id",
                                    "postal_code",
                                    "phone_number",
                                    "type"
                                ],
                                "properties": {
                                    "_method": {
                                        "description": "این مقدار باید بصورت ثابت شود",
                                        "type": "string",
                                        "default": "put",
                                        "enum": [
                                            "put"
                                        ]
                                    },
                                    "city_id": {
                                        "description": "شناسه شهر",
                                        "type": "number"
                                    },
                                    "address": {
                                        "description": "آدرس",
                                        "type": "string"
                                    },
                                    "postal_code": {
                                        "description": "کد پستی. بدون خط تیره مثال: 5717858731",
                                        "type": "number"
                                    },
                                    "phone_number": {
                                        "description": "شماره تلفن یا شماره موبایل. تفلن ثابت باید با کد شهر وارد شود بدون خط تیره مثال: 02162999977",
                                        "type": "number"
                                    },
                                    "type": {
                                        "description": "Home: خانه - Office: دفتر",
                                        "type": "string",
                                        "enum": [
                                            "Home",
                                            "Office"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "عملیات موفق",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "آدرس"
                ],
                "summary": "حذف آدرس",
                "description": "",
                "operationId": "88b5a2bf3ca4f7bb44c84127a6e4c119",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "شناسه آدرس",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "عملیات موفق",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "توکن را به همراه Bearer وارد کنید.",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}